+++ CHANGES.rst
@@ -13,6 +13,7 @@ Changelog
 * Allow destructive unittests in Travis CI jobs.
 * Allow usage of ``pip2`` and ``pip3`` managers on Linux.
 * Print current platform in debug messages.
+* Unittest detection of managers on each platform.
 
 
 `2.1.1 (2016-12-17) <https://github.com/kdeldycke/meta-package-manager/compare/v2.1.0...v2.1.1>`_